+++ cmd/hidden_lake/template/internal/config/init_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/settings"
 )
 
 const (