+++ cmd/hidden_lake/template/internal/handler/config_settings.go
@@ -3,8 +3,8 @@ package handler
 import (
 	"net/http"
 
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/internal/config"
 	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/internal/config"
 	"github.com/number571/go-peer/internal/api"
 	http_logger "github.com/number571/go-peer/internal/logger/http"
 	"github.com/number571/go-peer/pkg/logger"