+++ cmd/hidden_lake/template/internal/handler/config_settings_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/client"
+	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/client"
 	testutils "github.com/number571/go-peer/test/utils"
 )