+++ cmd/hidden_lake/template/internal/handler/index_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
-	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/client"
-	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
+	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/client"
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/settings"
 	testutils "github.com/number571/go-peer/test/utils"
 )