+++ cmd/hidden_lake/template/pkg/app/app.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"sync"
 
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/internal/config"
-	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/internal/config"
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/settings"
 	"github.com/number571/go-peer/internal/closer"
 	http_logger "github.com/number571/go-peer/internal/logger/http"
 	std_logger "github.com/number571/go-peer/internal/logger/std"