+++ cmd/hidden_lake/template/pkg/app/app_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/internal/config"
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/client"
-	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/internal/config"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/pkg/client"
+	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/template/pkg/settings"
 	testutils "github.com/number571/go-peer/test/utils"
 )