+++ cmd/hidden_lake/template/pkg/app/init_app.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/internal/config"
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/internal/config"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/pkg/settings"
 	"github.com/number571/go-peer/internal/flag"
 	"github.com/number571/go-peer/pkg/types"
 )