+++ cmd/hidden_lake/template/pkg/client/client.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/config"
+	"github.com/number571/go-peer/cmd/hidden_lake/template/pkg/config"
 )
 
 var (