+++ cmd/hidden_lake/template/pkg/config/types.go
@@ -1,6 +1,6 @@
 package config
 
-import "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/internal/config"
+import "github.com/number571/go-peer/cmd/hidden_lake/template/internal/config"
 
 type IConfigSettings config.IConfigSettings
 type SConfigSettings struct {